Revert "feat: Display Components Implementation"
