+++ README.md
@@ -197,16 +197,6 @@ fun App() {
 | **TabRow**               | Row of tabs                           | `selectedTabIndex`, `children`, `containerColor`, `contentColor`                                                                                                                              |
 | **ScrollableTabRow**     | Scrollable row of tabs                | `selectedTabIndex`, `children`, `containerColor`, `contentColor`                                                                                                                              |
 | **Tab**                  | Item for TabRow                       | `selected`, `text`, `icon`, `enabled`, `onClickEventName`                                                                                                                                     |
-| **Badge**                 | Display notification badge            | `text`, `textStateHostName`, `containerColor`, `contentColor`                                                                                                                                 |
-| **BadgedBox**             | Layout container for a badge & child  | `badge`, `child`                                                                                                                                                                             |
-| **AssistChip**            | Action chip                           | `label`, `leadingIcon`, `onClickEventName`, `enabled`, `enabledStateHostName`                                                                                                                 |
-| **SuggestionChip**        | Suggestion chip                       | `label`, `leadingIcon`, `onClickEventName`, `enabled`, `enabledStateHostName`                                                                                                                 |
-| **FilterChip**            | Selection filter chip                 | `selected`, `selectedStateHostName`, `label`, `leadingIcon`, `onClickEventName`, `enabled`, `enabledStateHostName`                                                                            |
-| **InputChip**             | Entry / tag chip                      | `selected`, `selectedStateHostName`, `label`, `leadingIcon`, `trailingIcon`, `onClickEventName`, `enabled`, `enabledStateHostName`                                                            |
-| **CircularProgressIndicator**| Circular progress indicator        | `progress`, `progressStateHostName`, `color`, `trackColor`                                                                                                                                    |
-| **LinearProgressIndicator**| Linear progress indicator            | `progress`, `progressStateHostName`, `color`, `trackColor`                                                                                                                                    |
-| **PlainTooltip**          | Plain tooltip                         | `text`, `anchor`                                                                                                                                                                              |
-| **RichTooltip**           | Rich tooltip                          | `title`, `text`, `action`, `anchor`                                                                                                                                                          |
 | **Custom**               | User-defined components               | `customType`, `customData`                                                                                                                                                                    |
 | **Switch**               | Material toggle switch                | `checked`, `onCheckedChange`, `enabled`                                                                                                                                                      |
 | **Checkbox**             | Material checkbox                     | `checked`, `onCheckedChange`, `enabled`                                                                                                                                                      |
